+++ tests/jsonpickle_test.py
@@ -75,7 +75,6 @@ class ThingWithProps(object):
         return self.identity == other.identity
 
 
-
 class PicklingTestCase(unittest.TestCase):
 
     def setUp(self):
@@ -161,17 +160,17 @@ class PicklingTestCase(unittest.TestCase):
         self.assertEqual(tupleB, self.unpickler.restore(tupleB_pickle))
 
     def test_tuple_roundtrip(self):
-        data = (1,2,3)
+        data = (1, 2, 3)
         newdata = jsonpickle.decode(jsonpickle.encode(data))
         self.assertEqual(data, newdata)
 
     def test_set_roundtrip(self):
-        data = set([1,2,3])
+        data = set([1, 2, 3])
         newdata = jsonpickle.decode(jsonpickle.encode(data))
         self.assertEqual(data, newdata)
 
     def test_list_roundtrip(self):
-        data = [1,2,3]
+        data = [1, 2, 3]
         newdata = jsonpickle.decode(jsonpickle.encode(data))
         self.assertEqual(data, newdata)
 
@@ -206,7 +205,7 @@ class PicklingTestCase(unittest.TestCase):
         self.assertEqual('a string', inflated[2])
 
     def test_classdict(self):
-        dict = {'k1':Thing('one'), 'k2':Thing('two'), 'k3':3}
+        dict = {'k1': Thing('one'), 'k2': Thing('two'), 'k3': 3}
 
         flattened = self.pickler.flatten(dict)
         self.assertEqual('one', flattened['k1']['name'])
@@ -300,9 +299,8 @@ class PicklingTestCase(unittest.TestCase):
         obj.typeref = object
 
         flattened = self.pickler.flatten(obj)
-        self.assertEqual(flattened['typeref'], {
-                            tags.TYPE: '__builtin__.object',
-                         })
+        self.assertEqual(flattened['typeref'],
+                         {tags.TYPE: '__builtin__.object'})
 
         inflated = self.unpickler.restore(flattened)
         self.assertEqual(inflated.typeref, object)
@@ -316,9 +314,8 @@ class PicklingTestCase(unittest.TestCase):
         obj.classref = Thing
 
         flattened = self.pickler.flatten(obj)
-        self.assertEqual(flattened['classref'], {
-                            tags.TYPE: 'jsonpickle_test.Thing',
-                         })
+        self.assertEqual(flattened['classref'],
+                         {tags.TYPE: 'jsonpickle_test.Thing'})
 
         inflated = self.unpickler.restore(flattened)
         self.assertEqual(inflated.classref, Thing)
@@ -379,8 +376,9 @@ class JSONPickleTestCase(unittest.TestCase):
     def setUp(self):
         self.obj = Thing('A name')
         self.expected_json = (
-                '{"'+tags.OBJECT+'": "jsonpickle_test.Thing",'
-                ' "name": "A name", "child": null}')
+            '{"%s": "jsonpickle_test.Thing", "name": "A name", "child": null}'
+            % tags.OBJECT
+        )
 
     def test_encode(self):
         expect = self.obj
@@ -420,7 +418,7 @@ class JSONPickleTestCase(unittest.TestCase):
 
     def test_tuple_dict_keys_default(self):
         """Test that we handle dictionaries with tuples as keys."""
-        tuple_dict = {(1, 2): 3, (4, 5): { (7, 8): 9 }}
+        tuple_dict = {(1, 2): 3, (4, 5): {(7, 8): 9}}
         pickle = jsonpickle.encode(tuple_dict)
         expect = {'(1, 2)': 3, '(4, 5)': {'(7, 8)': 9}}
         actual = jsonpickle.decode(pickle)
@@ -433,7 +431,7 @@ class JSONPickleTestCase(unittest.TestCase):
 
     def test_tuple_dict_keys_with_keys_enabled(self):
         """Test that we handle dictionaries with tuples as keys."""
-        tuple_dict = {(1, 2): 3, (4, 5): { (7, 8): 9 }}
+        tuple_dict = {(1, 2): 3, (4, 5): {(7, 8): 9}}
         pickle = jsonpickle.encode(tuple_dict, keys=True)
         expect = tuple_dict
         actual = jsonpickle.decode(pickle, keys=True)
@@ -702,7 +700,7 @@ class PicklableNamedTupleEx(object):
         ntuple.__getnewargs_ex__ = (lambda self: ((), kwargs))
         ntuple.__getnewargs__ = newargs
         instance = ntuple.__new__(ntuple,
-                                  *[b for a,b in sorted(kwargs.items())])
+                                  *[b for a, b in sorted(kwargs.items())])
         return instance
 
 
@@ -734,6 +732,7 @@ class PickleProtocol2Thing(object):
 slotmagic = PickleProtocol2Thing('slotmagic')
 dictmagic = PickleProtocol2Thing('dictmagic')
 
+
 class PickleProtocol2GetState(PickleProtocol2Thing):
     def __new__(cls, *args):
         instance = super(PickleProtocol2GetState, cls).__new__(cls)
@@ -743,13 +742,16 @@ class PickleProtocol2GetState(PickleProtocol2Thing):
     def __getstate__(self):
         return 'I am magic'
 
+
 class PickleProtocol2GetStateDict(PickleProtocol2Thing):
     def __getstate__(self):
         return {'magic': True}
 
+
 class PickleProtocol2GetStateSlots(PickleProtocol2Thing):
     def __getstate__(self):
         return (None, {'slotmagic': slotmagic})
+
 
 class PickleProtocol2GetStateSlotsDict(PickleProtocol2Thing):
     def __getstate__(self):
@@ -800,7 +802,7 @@ class PickleProtocol2ReduceTuple(object):
                 ('yam', 1),  # args
                 None,  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -812,12 +814,13 @@ class PickleProtocol2ReduceTupleFunc(object):
     def __init__(self, argval, optional=None):
         self.argval = argval
         self.optional = optional
+
     def __reduce__(self):
         return (protocol_2_reduce_tuple_func,  # callable
                 ('yam', 1),  # args
                 None,  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -839,7 +842,7 @@ class PickleProtocol2ReduceNewobj(PickleProtocol2ReduceTupleFunc):
                 (PickleProtocol2ReduceNewobj, 'yam', 1),  # args
                 None,  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -849,7 +852,7 @@ class PickleProtocol2ReduceTupleState(PickleProtocol2ReduceTuple):
                 ('yam', 1),  # args
                 {'foo': 1},  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -862,7 +865,7 @@ class PickleProtocol2ReduceTupleSetState(PickleProtocol2ReduceTuple):
                 ('yam', 1),  # args
                 {'foo': 1},  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -878,7 +881,7 @@ class PickleProtocol2ReduceTupleStateSlots(object):
                 ('yam', 1),  # args
                 {'foo': 1},  # state
                 iter([]),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
 
@@ -891,11 +894,12 @@ class PickleProtocol2ReduceListitemsAppend(object):
                 (),  # args
                 {},  # state
                 iter(['foo', 'bar']),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
     def append(self, item):
         self.inner.append(item)
+
 
 class PickleProtocol2ReduceListitemsExtend(object):
     def __init__(self):
@@ -906,11 +910,12 @@ class PickleProtocol2ReduceListitemsExtend(object):
                 (),  # args
                 {},  # state
                 iter(['foo', 'bar']),  # listitems
-                iter([]), # dictitems
+                iter([]),  # dictitems
                 )
 
     def extend(self, items):
         self.inner.exend(items)
+
 
 class PickleProtocol2ReduceDictitems(object):
     def __init__(self):
@@ -921,11 +926,12 @@ class PickleProtocol2ReduceDictitems(object):
                 (),  # args
                 {},  # state
                 [],  # listitems
-                iter(zip(['foo', 'bar'],['foo', 'bar'])), # dictitems
+                iter(zip(['foo', 'bar'], ['foo', 'bar'])),  # dictitems
                 )
 
     def __setitem__(self, k, v):
         return self.inner.__setitem__(k, v)
+
 
 class PickleProtocol2Classic:
 
@@ -938,7 +944,7 @@ class PickleProtocol2ClassicInitargs:
     def __init__(self, foo, bar=None):
         self.foo = foo
         if bar:
-            self.bar=bar
+            self.bar = bar
 
     def __getinitargs__(self):
         return ('choo', 'choo')
@@ -965,7 +971,7 @@ class PicklingProtocol4TestCase(unittest.TestCase):
         instance = PicklableNamedTupleEx(**{'a': 'b', 'n': 2})
         args, kwargs = instance.__getnewargs_ex__()
         newinstance = PicklableNamedTupleEx.__new__(PicklableNamedTupleEx,
-                                                  *args, **kwargs)
+                                                    *args, **kwargs)
         self.assertEqual(instance, newinstance)
 
     def test_references(self):
@@ -1056,9 +1062,8 @@ class PicklingProtocol2TestCase(unittest.TestCase):
         self.assertEqual(decoded.inner, ['foo', 'bar'])
 
     def test_reduce_state_setstate(self):
-        'Test reduce with the optional state argument set, on an object with a'\
-        '__setstate__'
-        # nosetests only shows first line of docstring
+        'Test reduce with the optional state argument set, on an object with '\
+            'a __setstate__'
 
         instance = PickleProtocol2ReduceTupleSetState(5)
         encoded = jsonpickle.encode(instance)
@@ -1069,8 +1074,8 @@ class PicklingProtocol2TestCase(unittest.TestCase):
         self.assertFalse(hasattr(decoded, 'foo'))
 
     def test_reduce_state_no_dict(self):
-        'Test reduce with the optional state argument set, on an object with'\
-        'no __dict__, and no __setstate__'
+        'Test reduce with the optional state argument set, on an object with '\
+            'no __dict__, and no __setstate__'
 
         instance = PickleProtocol2ReduceTupleStateSlots(5)
         encoded = jsonpickle.encode(instance)
@@ -1080,8 +1085,8 @@ class PicklingProtocol2TestCase(unittest.TestCase):
         self.assertEqual(decoded.foo, 1)
 
     def test_reduce_state_dict(self):
-        'Test reduce with the optional state argument set, on an object with a'\
-        '__dict__, and no __setstate__'
+        'Test reduce with the optional state argument set, on an object with '\
+            'a __dict__, and no __setstate__'
 
         instance = PickleProtocol2ReduceTupleState(5)
         encoded = jsonpickle.encode(instance)
@@ -1150,7 +1155,6 @@ class PicklingProtocol2TestCase(unittest.TestCase):
         decoded = jsonpickle.decode(encoded)
         self.assertEqual(decoded, slotmagic)
 
-
     def test_pickle_newargs(self):
         """
         Ensure we can pickle and unpickle an object whose class needs arguments
@@ -1169,7 +1173,7 @@ class PicklingProtocol2TestCase(unittest.TestCase):
         """
         instance = PicklableNamedTuple(('a', 'b'), (1, 2))
         newinstance = PicklableNamedTuple.__new__(PicklableNamedTuple,
-                                                 *(instance.__getnewargs__()))
+                                                  *(instance.__getnewargs__()))
         self.assertEqual(instance, newinstance)
 
     def test_getnewargs_priority(self):
@@ -1245,7 +1249,7 @@ class PicklingProtocol2TestCase(unittest.TestCase):
     def test_handles_cyclical_objects(self):
         child = PickleProtocol2Thing(None)
         instance = PickleProtocol2Thing(child, child)
-        child.args = (instance,) # create a cycle
+        child.args = (instance,)  # create a cycle
         # TODO we do not properly restore references inside of lists.
         # Change the above tuple into a list to show the breakage.
 
@@ -1278,7 +1282,7 @@ class PicklingProtocol2TestCase(unittest.TestCase):
     def test_handles_cyclical_objects_in_lists(self):
         child = PickleProtocol2ChildThing(None)
         instance = PickleProtocol2ChildThing([child, child])
-        child.child = instance # create a cycle
+        child.child = instance  # create a cycle
 
         encoded = jsonpickle.encode(instance)
         decoded = jsonpickle.decode(encoded)