+++ docs/source/changelog.rst
@@ -11,6 +11,8 @@ Version 0.8.1 - January TBD, 2015
       and allow custom handlers to register themselves for all subclasses.
       (`#104 <https://github.com/jsonpickle/jsonpickle/pull/104>`_).
 
+    * We now support serializing Python3 `Enum` classes.
+
 Version 0.8.0 - September 6, 2014
 ---------------------------------